Move account creation and consent uploading into separate steps
To upload consent results, the user must be logged in.
The consent object is still validated and will throw an
error if it's invalid. If the current user does not have
a family and given name, it is extracted and updated. Otherwise
the family and given name are still saved on the signature object,
but otherwise ignored. This closes #19.