Restrict save/fetch methods to ORKTaskResult (and subclass) instances
Initially, the SDK supported the serialization and saving of any ORKResult
or subclass. ORKTaskViewController is the primary entry point for the
ResearchKit UI Framework. As such, it is very unlikely would need to
save any ORKResult that was not:

 1. An ORKTaskResult
 2. Nested inside an ORKTaskResult

Restricting the save and fetch methods has advantages. Namely,
ORKTaskResult instances have a taskRunUUID property. This will
allow us to use the taskRunUUID as the CMObject objectId,
thereby seamlessy enabling developers to restore and update
existing ORKTaskResult data already saved to CloudMine in a
way that is fully compatible with the design of ResearchKit.