Prevent fetch from starting before a push is enqueued
The sync queue was implemented such that all push operations
had to finish before a fetch operation could execute. However,
the push operation is queued only after it is successfully
persisted in the store. It was possible for a fetch call to occur
*after* an update had persisted, but *before* said update was
enqueued to push to the server. Thus stale data could be fetched,
and updated data could be ovewritten locally, even while being
pushed to the server.

The resolution to this issue is a bit ugly, but seems to work. When
an update is sent to the store, we increment a pre-queue counter on
the sync object. Fetch operations are delayed until this counter is
0. When the update is enqueued or fails local persistence, we decrement
the pre-queue counter.