Adds full blown NSOpertion subclass for the care pull operation
Previously, sync operations were simply performed in a block
that was added directly to an internal operation queue. This worked,
but as a result, sync operations could sometimes *finish* in a
different order than they were called.

This is beause the block itself, with internal network operations,
would take an indeterminate time to execute. So though fetch
requests would *start* in the order they were called, they might
finish out of order, because the Operation was considered "done"
as soon as the block was called.

By implementing a full blown NSOperation subclass, which does not
finish until the fetch operation itself is complete, we can
serialize the fetch operations and guarentee they start *and*
finish in the order called.