Handle OCKActivity modify-by-delete-then-replace properly
Most properties on an Activity cannot be edited. To "change"
and activity, the framework encourages removing the activity from
the local store, creating a modified version with the same
identifier, and adding it back in.

This creates a problem on the backend. Because we need to persist even
"deleted" data, we simply _flag_ an activity as deleted, rather than
actually removing it. However, if a new activity is added with the
same identifier, the new activity will collide with the one which is
still stored, but flagged as deleted. That is because the CloudMine
object id is derived from the activity identifier.

To resolve this issue, we take a two step process to deletion:

 1. Create an archived, flagged-as-deleted copy of the Activity, with a
    a `__DELETED__` delimeter and a UUID appended to the original object Id
 2. Delete the original representation of the Activity on CloudMine

To implement step two, we added a new delete operation to the internal queue.
During the synchronization process, it is now neccessary to seperate delete
operations from add/modify operations, and ensure the former are applied first.
Furthermore, during activity syncrhonization, the process must now check the
current state of the store before each operation, as operations can change
the state of the store, and thus the action the synchronization process must
take.

A major issue remains with this approach. The deletion of an activity, and
subsequent addition of a new activity with the same identifier, may leave
events stored that are now invalid. For example, if an activity was previously
twice a day and now is only once a day. These events, when the SDK attempts
to synchronize them, will lead to crashes. There is no easy fix for this.
Right now, the only viable approach would be to archive all Events for
a given activity at the same time that the activity is archived.