Prevent Syncing Invalid Events
When an Activity is deleted, it is possible for another
activity to be added with the same identifier as the one
previously deleted. We archive the old activity on the backend,
but Events associated with the old Activity could still be
fetched and added to the local store, associating with the
new Activity. This is incorrect.

Now, when an Event is synced, it first compare its Activity to
the one in the local store with the same identifier. If the
Activity is not present, or is not identical (except for end
date, which can be mutated in place), then the Event is not
added to the local store.